# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with BookStore Application",
        "description": "Integrate Swagger in Laravel  BookStoreapplication",
        "contact": {
            "email": "chandanmohanty11111@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/addbook": {
            "post": {
                "summary": "Add Book",
                "description": "Admin Can Add Book ",
                "operationId": "ec4e9b9f7a7c56771d9143574c257748",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "author",
                                    "image",
                                    "Price",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "Price": {
                                        "type": "decimal"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/updatebook": {
            "post": {
                "summary": "Update Book",
                "description": "Admin Can Update Book ",
                "operationId": "3f74d6bed68a22115f57e6cc1f83e1ee",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "description",
                                    "author",
                                    "image",
                                    "Price"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "Price": {
                                        "type": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book updated Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/addquantity": {
            "post": {
                "summary": "Add Quantity to Existing Book",
                "description": " Add Book Quantity ",
                "operationId": "52974593a183cab8cc4c6554b361b255",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "quantity"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Quantity updated Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/deletebook": {
            "post": {
                "summary": "Delete the book from BookStoreApp",
                "description": " Delete Book ",
                "operationId": "68eaebcbf2fba7119ae4fe3c29a60487",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book deleted Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/displaybooks": {
            "get": {
                "summary": "Display All Books",
                "description": " Display All Books Present in the BookStore ",
                "operationId": "42ecfb3cd87d86dd01400517f00db529",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Books Available in the Bookstore are"
                    },
                    "404": {
                        "description": "Books are not there"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}